Qt Project files for buildTree
Started to test whether tree-building algorithm is correct. Right now, running the main.cpp will print to terminal the value of the node that is added to the tree. Right now, the outermost for loop is hard-coded to 1 because I just wanted to see the first iteration. Should be for the number of data points in the sample file.